// frontend/src/api/generated/model/criterionAnswerType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Call Analyzer Api
 * OpenAPI spec version: 0.1.0
 */

export type CriterionAnswerType = typeof CriterionAnswerType[keyof typeof CriterionAnswerType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CriterionAnswerType = {
  text: 'text',
  percentage: 'percentage',
  boolean: 'boolean',
} as const;