// frontend/src/api/generated/model/transcriptionStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Call Analyzer Api
 * OpenAPI spec version: 0.1.0
 */

export type TranscriptionStatus = typeof TranscriptionStatus[keyof typeof TranscriptionStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TranscriptionStatus = {
  uploaded: 'uploaded',
  queued: 'queued',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;